fix: change files names and fixes some undefined objet - Ref gestion-de-projet#2027
