+++ src/components/ScopeTree/index.tsx
@@ -1,12 +1,11 @@
-import React from 'react'
-import { useState } from 'react'
+import React, { useState } from 'react'
 import { useAppSelector } from 'state'
 import { ScopeState } from 'state/scope'
 import { ScopeTreeRow, ScopeType } from 'types'
 import ScopeSearchBar from '../Inputs/ScopeSearchBar/ScopeSearchBar'
-import ScopeTreeChipsets from './ScopeTreeChipsets/ScopeTreeChipsets'
+import ScopeTreeChipsets from './ScopeTreeChipsets'
 import ScopeTreeExploration from './ScopeTreeExploration'
-import ScopeTreeSearch from './ScopeTreeSearch/./index'
+import ScopeTreeSearch from './ScopeTreeSearch'
 import useStyles from './styles'
 import { onSelect } from './utils/scopeTreeUtils'
